Moved DiagrammedExpr.scala into org.scalatest.diagrams package for dotty build.
