+++ dotty/diagrams/src/main/scala/org/scalatest/diagrams/DiagramsMacro.scala
@@ -17,7 +17,7 @@ package org.scalatest.diagrams
 
 import org.scalactic._
 import scala.quoted._
-import org.scalatest.{Assertions, DiagrammedExpr}
+import org.scalatest.Assertions
 import org.scalatest.compatible.Assertion
 
 object DiagramsMacro {
